MakeGeneric
Made the model baseclass seperate from the other sections
