--- ppLFER_MUM/debugger.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Wed Jul 25 10:07:02 2018
-
-@author: Tim Rodgers
-"""
-
